+++ packages/appcd-client/CHANGELOG.md
@@ -1,5 +1,6 @@
 # v2.0.5
 
+ * feat: Added `startDaemon` flag to `request()` method which passes it through to `connect()`.
  * chore: Updated dependencies.
 
 # v2.0.4 (Jan 13, 2020)